requirements.txt
pandas
openpyxl
selenium
python-dotenv
langgraph 
langsmith 
langchain  
langchain_community
sqlalchemy 
google-generativeai
langchain-google-genai
streamlit
streamlit-option-menu
python-pptx 
pymupdf 
pytesseract
chromadb
langchain-chroma
pypdf